created new types
