--- src/utils/nodeMailer.js
@@ -1,20 +1,0 @@
-const nodemailer = require('nodemailer')
-
-const transporter = nodemailer.createTransport({
-  host: 'gmail',
-  auth: {
-    user: process.env.AUTH_EMAIL,
-    pass: process.env.AUTH_PASS
-  }
-})
-
-transporter.verify((error, success) => {
-  if (error) {
-    console.log(error)
-  } else {
-    console.log('Ready for message')
-    console.log(success)
-  }
-})
-
-module.exports = transporter
